refactor: fix resources import
