+++ core/ui/resources/src/main/res/drawable/f1_logo.xml
