+++ core/ui/resources/src/main/res/drawable/ic_schedule.xml
